skill-assessment-agent/requirements.txt
streamlit==1.40.0
openai==1.55.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pydantic==2.10.0
plotly==5.24.0
pandas
